finishing product

+++ app/Http/Livewire/Product/DeleteProduct.php
@@ -4,30 +4,40 @@ namespace App\Http\Livewire\Product;
 
 use App\Models\Product;
 use LivewireUI\Modal\ModalComponent;
+use Illuminate\Support\Facades\Storage;
 
 class DeleteProduct extends ModalComponent
 {
     public $productId;
+    public $photo;
 
     public function mount(Product $product)
     {
         $this->productId = $product; 
+        $this->photo = $product->photo;
     }
     public function deleteProduct()
     {
-        $this->productId->delete();
-        $this->closeModalWithEvents(['productsUpdated',
-        IndexProduct::getName() => 'productsUpdated',]);
-        $this->alert('success', 'Deleted!', [
-            'position' =>  'top-end', 
-            'timer' =>  '4000', 
-            'toast' =>  true, 
-            'text' =>  '', 
-            'confirmButtonText' =>  'Ok', 
-            'cancelButtonText' =>  'Cancel', 
-            'showCancelButton' =>  false, 
-            'showConfirmButton' =>  false, 
-      ]);
+        if($this->productId)
+        {
+            if(Storage::exists("public/$this->photo"))
+            {
+                $this->productId->delete();
+                Storage::delete("public/$this->photo");
+                $this->closeModalWithEvents(['productsUpdated',
+                IndexProduct::getName() => 'productsUpdated',]);
+                $this->alert('success', 'Deleted!', [
+                    'position' =>  'top-end', 
+                    'timer' =>  '4000', 
+                    'toast' =>  true, 
+                    'text' =>  '', 
+                    'confirmButtonText' =>  'Ok', 
+                    'cancelButtonText' =>  'Cancel', 
+                    'showCancelButton' =>  false, 
+                    'showConfirmButton' =>  false, 
+                ]);
+            }
+        }
     }
 
     public function render()
